Comiit

+++ src/app/components/login/login.component.ts
@@ -21,37 +21,43 @@ export class LoginComponent implements OnInit {
   ngOnInit() {
   }
 
-
   login(){
-    var params = { 
-      cedula: this.cedula,
-      contrasena: this.password
-     }
-    this.genericService.autenticar(this.componentUrl+"autenticar", params, (estudiante) => {
-      if(estudiante){
-        localStorage.setItem('matricula', estudiante.matricula);
-        localStorage.setItem('cedula', estudiante.cedula);
-        localStorage.setItem('nivelId', estudiante.nivel.id);
-        localStorage.setItem('institucionId', estudiante.institucion.id);
-        this._router.navigate(['/solicitudes']);
-      }
-      else{
-        Swal.fire({
-          title: "Usuario no encontrado",
-          text: "Desea crear un nuevo usuario",
-          type: "info",
-          showCancelButton: true,
-          cancelButtonText: "No",
-          confirmButtonText: "Sí"
-        }).then(response => {
-          if(response.value){
-            this._router.navigate(['/nuevo']);
-          }
-        })
+    //Luego de la mascara, el string llegaba sin los dashes (-), aqui se los agregamos.
+    this.cedula = `${this.cedula.substr(0,3)}-${this.cedula.substr(3,7)}-${this.cedula.substr(10,1)}`
+    if(this.password == ""){
+      Swal.fire("Contraseña incorrecta", "Debe introducir una contraseña válida", "warning" );
+    }
+    else{
+      var params = { 
+        cedula: this.cedula,
+        contrasena: this.password
+       }
+      this.genericService.autenticar(this.componentUrl+"autenticar", params, (estudiante) => {
+        if(estudiante){
+          localStorage.setItem('matricula', estudiante.matricula);
+          localStorage.setItem('cedula', estudiante.cedula);
+          localStorage.setItem('nivelId', estudiante.nivel.id);
+          localStorage.setItem('institucionId', estudiante.institucion.id);
+          this._router.navigate(['/solicitudes']);
+        }
+        else{
+          Swal.fire({
+            title: "Usuario no encontrado",
+            text: "Desea crear un nuevo usuario",
+            type: "info",
+            showCancelButton: true,
+            cancelButtonText: "No",
+            confirmButtonText: "Sí"
+          }).then(response => {
+            if(response.value){
+              this._router.navigate(['/nuevo']);
+            }
+          })
+          
+        }
         
-      }
-      
-    })
+      })
+    }
   }
 
 }
